+++ src/index.ts
@@ -59,15 +59,22 @@ import {
     AlgorithmStorageKey,
     BooleanFilterName,
     CacheKey,
+    FediverseCacheKey,
+    LoadAction,
+    LogAction,
     MediaCategory,
     NonScoreWeightName,
     ScoreName,
     TrendingType,
     TypeFilterName,
     TagTootsCategory,
+    ALL_ACTIONS,
     JUST_MUTING,
+    // LOG_KEYS,
     buildCacheKeyDict,
     isValueInStringEnum,
+    type Action,
+    type ApiCacheKey,
 } from "./enums";
 import {
     computeMinMax,
@@ -98,29 +105,6 @@ import {
     type Weights,
 } from "./types";
 
-enum LogPrefix {
-    FINISH_FEED_UPDATE = 'finishFeedUpdate',
-    REFRESH_MUTED_ACCOUNTS = 'refreshMutedAccounts',
-    RESET = 'reset',
-    TRIGGER_FEED_UPDATE = "triggerFeedUpdate",
-    TRIGGER_MOAR_DATA = 'triggerMoarData',
-    TRIGGER_PULL_ALL_USER_DATA = "triggerPullAllUserData",
-    TRIGGER_TIMELINE_BACKFILL = "triggerTimelineBackfill",
-};
-
-const LOADING_STATUS_MSGS: {[key in LogPrefix]?: string} = {
-    [LogPrefix.RESET]: `Resetting state`,
-    [LogPrefix.TRIGGER_MOAR_DATA]: `Fetching more data for the algorithm`,
-    [LogPrefix.TRIGGER_PULL_ALL_USER_DATA]: `Pulling your historical data`,
-    [LogPrefix.TRIGGER_TIMELINE_BACKFILL]: `Loading older home timeline toots`,
-};
-
-const FINALIZING_SCORES_MSG = `Finalizing scores`;
-const GET_FEED_BUSY_MSG = `Load in progress (consider using the setTimelineInApp() callback instead)`;
-const INITIAL_LOAD_STATUS = "Retrieving initial data";
-const READY_TO_LOAD_MSG = "Ready to load";
-const DEFAULT_SET_TIMELINE_IN_APP = (_feed: Toot[]) => console.debug(`Default setTimelineInApp() called`);
-
 const EMPTY_TRENDING_DATA: TrendingData = {
     links: [],
     tags: new TagList([], TagTootsCategory.TRENDING),
@@ -128,9 +112,20 @@ const EMPTY_TRENDING_DATA: TrendingData = {
     toots: []
 };
 
+const DEFAULT_SET_TIMELINE_IN_APP = (_feed: Toot[]) => console.debug(`Default setTimelineInApp() called`);
+
 const logger = new Logger(`TheAlgorithm`);
-const loggers: Record<string, Logger> = buildCacheKeyDict((key) => new Logger(key as string));
-Object.values(LogPrefix).forEach((prefix) => loggers[prefix] = logger.tempLogger(prefix));
+
+const loggers: Record<Action | ApiCacheKey, Logger> = buildCacheKeyDict<Action, Logger, Record<Action, Logger>>(
+    (key) => new Logger(key as string),
+    ALL_ACTIONS.reduce(
+        (_loggers, action) => {
+            _loggers[action] = logger.tempLogger(action);
+            return _loggers;
+        },
+        {} as Record<Action, Logger>
+    )
+);
 
 interface AlgorithmArgs {
     api: mastodon.rest.Client;
@@ -164,7 +159,7 @@ interface AlgorithmArgs {
  * @property {UserData} userData - User data for scoring and filtering
  * @property {WeightInfoDict} weightsInfo - Info about all scoring weights
  */
-class TheAlgorithm {
+export default class TheAlgorithm {
     /**
      * True if FEDIALGO_DEBUG environment var was set at compile time.
      * @returns {boolean}
@@ -178,7 +173,7 @@ class TheAlgorithm {
 
     filters: FeedFilterSettings = buildNewFilterSettings();
     lastLoadTimeInSeconds?: number;
-    loadingStatus: string | null = READY_TO_LOAD_MSG;
+    loadingStatus: string | null = config.locale.messages[LogAction.INITIAL_LOADING_STATUS];
     trendingData: TrendingData = EMPTY_TRENDING_DATA;
 
     get apiErrorMsgs(): string[] { return MastoApi.instance.apiErrors.map(e => e.message) };
@@ -286,16 +281,17 @@ class TheAlgorithm {
 
     /**
      * Trigger the retrieval of the user's timeline from all the sources.
-     * @param {boolean} [moreOldToots] - Backfill older toots instead of getting new toots
      * @returns {Promise<void>}
      */
     async triggerFeedUpdate(): Promise<void> {
         if (this.shouldSkip()) return;
-        await this.lockLoadingMutex(LogPrefix.TRIGGER_FEED_UPDATE);
+        const action = LoadAction.FEED_UPDATE;
+        const hereLogger = loggers[action];
+        await this.startAction(action);
 
         try {
             const tootsForHashtags = async (key: TagTootsCategory): Promise<Toot[]> => {
-                loggers[LogPrefix.TRIGGER_FEED_UPDATE].trace(`Fetching toots for hashtags with key: ${key}`);
+                hereLogger.trace(`Fetching toots for hashtags with key: ${key}`);
                 const tagList = await TagsForFetchingToots.create(key);
                 return await this.fetchAndMergeToots(tagList.getToots(), tagList.logger);
             };
@@ -304,7 +300,7 @@ class TheAlgorithm {
                 // Toot fetchers
                 this.getHomeTimeline().then((toots) => this.homeFeed = toots),
                 this.fetchAndMergeToots(MastoApi.instance.getHomeserverToots(), loggers[CacheKey.HOMESERVER_TOOTS]),
-                this.fetchAndMergeToots(MastodonServer.fediverseTrendingToots(), loggers[CacheKey.FEDIVERSE_TRENDING_TOOTS]),
+                this.fetchAndMergeToots(MastodonServer.fediverseTrendingToots(), loggers[FediverseCacheKey.FEDIVERSE_TRENDING_TOOTS]),
                 ...Object.values(TagTootsCategory).map(async (key) => await tootsForHashtags(key)),
                 // Other data fetchers
                 MastodonServer.getTrendingData().then((trendingData) => this.trendingData = trendingData),
@@ -313,10 +309,10 @@ class TheAlgorithm {
             ];
 
             const allResults = await Promise.allSettled(dataLoads);
-            loggers[LogPrefix.TRIGGER_FEED_UPDATE].deep(`FINISHED promises, allResults:`, allResults);
+            hereLogger.deep(`FINISHED promises, allResults:`, allResults);
             await this.finishFeedUpdate();
         } finally {
-            this.releaseLoadingMutex(LogPrefix.TRIGGER_FEED_UPDATE);
+            this.releaseLoadingMutex(action);
         }
     }
 
@@ -325,13 +321,13 @@ class TheAlgorithm {
      * @returns {Promise<void>}
      */
     async triggerHomeTimelineBackFill(): Promise<void> {
-        await this.lockLoadingMutex(LogPrefix.TRIGGER_TIMELINE_BACKFILL);
+        await this.startAction(LoadAction.TIMELINE_BACKFILL);
 
         try {
             this.homeFeed = await this.getHomeTimeline(true);
             await this.finishFeedUpdate();
         } finally {
-            this.releaseLoadingMutex(LogPrefix.TRIGGER_TIMELINE_BACKFILL);
+            this.releaseLoadingMutex(LoadAction.TIMELINE_BACKFILL);
         }
     }
 
@@ -341,7 +337,7 @@ class TheAlgorithm {
      * @returns {Promise<void>}
      */
     async triggerMoarData(): Promise<void> {
-        await this.lockLoadingMutex(LogPrefix.TRIGGER_MOAR_DATA);
+        await this.startAction(LoadAction.GET_MOAR_DATA);
         let shouldReenablePoller = false;
 
         try {
@@ -358,7 +354,7 @@ class TheAlgorithm {
             throwSanitizedRateLimitError(error, `triggerMoarData() Error pulling user data:`);
         } finally {
             if (shouldReenablePoller) this.enableMoarDataBackgroundPoller();  // Reenable poller when finished
-            this.releaseLoadingMutex(LogPrefix.TRIGGER_MOAR_DATA);
+            this.releaseLoadingMutex(LoadAction.GET_MOAR_DATA);
         }
     }
 
@@ -368,8 +364,9 @@ class TheAlgorithm {
      * @returns {Promise<void>}
      */
     async triggerPullAllUserData(): Promise<void> {
-        const hereLogger = loggers[LogPrefix.TRIGGER_PULL_ALL_USER_DATA];
-        this.lockLoadingMutex(LogPrefix.TRIGGER_PULL_ALL_USER_DATA);
+        const action = LoadAction.PULL_ALL_USER_DATA;
+        const hereLogger = loggers[action];
+        this.startAction(action);
 
         try {
             this.dataPoller && clearInterval(this.dataPoller!);   // Stop the dataPoller if it's running
@@ -385,7 +382,7 @@ class TheAlgorithm {
         } catch (error) {
             throwSanitizedRateLimitError(error, hereLogger.line(`Error pulling user data:`));
         } finally {
-            this.releaseLoadingMutex(LogPrefix.TRIGGER_PULL_ALL_USER_DATA);  // TODO: should we restart data poller?
+            this.releaseLoadingMutex(action);  // TODO: should we restart data poller?
         }
     }
 
@@ -454,13 +451,13 @@ class TheAlgorithm {
      * @returns {Promise<void>}
      */
     async refreshMutedAccounts(): Promise<void> {
-        const hereLogger = loggers[LogPrefix.REFRESH_MUTED_ACCOUNTS];
+        const hereLogger = loggers[LoadAction.REFRESH_MUTED_ACCOUNTS];
         hereLogger.log(`called (${Object.keys(this.userData.mutedAccounts).length} current muted accounts)...`);
         // TODO: move refreshMutedAccounts() to UserData class?
         const mutedAccounts = await MastoApi.instance.getMutedAccounts({bustCache: true});
         hereLogger.log(`Found ${mutedAccounts.length} muted accounts after refresh...`);
         this.userData.mutedAccounts = Account.buildAccountNames(mutedAccounts);
-        await Toot.completeToots(this.feed, loggers[LogPrefix.REFRESH_MUTED_ACCOUNTS], JUST_MUTING);
+        await Toot.completeToots(this.feed, hereLogger, JUST_MUTING);
         await this.finishFeedUpdate();
     }
 
@@ -470,7 +467,7 @@ class TheAlgorithm {
      * @returns {Promise<void>}
      */
     async reset(complete: boolean = false): Promise<void> {
-        await this.lockLoadingMutex(LogPrefix.RESET);
+        await this.startAction(LoadAction.RESET);
 
         try {
             this.dataPoller && clearInterval(this.dataPoller!);
@@ -478,9 +475,10 @@ class TheAlgorithm {
             this.cacheUpdater && clearInterval(this.cacheUpdater!);
             this.cacheUpdater = undefined;
             this.hasProvidedAnyTootsToClient = false;
-            this.loadingStatus = READY_TO_LOAD_MSG;
+            this.loadingStatus = config.locale.messages[LogAction.INITIAL_LOADING_STATUS];
             this.loadStartedAt = undefined;
             this.numTriggers = 0;
+            this.trendingData = EMPTY_TRENDING_DATA;
             this.feed = [];
             this.setTimelineInApp([]);
 
@@ -495,7 +493,7 @@ class TheAlgorithm {
                 await this.loadCachedData();
             }
         } finally {
-            this.releaseLoadingMutex(LogPrefix.RESET);
+            this.releaseLoadingMutex(LoadAction.RESET);
         }
     }
 
@@ -582,13 +580,14 @@ class TheAlgorithm {
 
     // Return true if we're in QUICK_MODE and the feed is fresh enough that we don't need to update it (for dev)
     private shouldSkip(): boolean {
-        loggers[LogPrefix.TRIGGER_FEED_UPDATE].info(`${++this.numTriggers} triggers so far, state:`, this.statusDict());
+        const hereLogger = loggers[LoadAction.FEED_UPDATE];
+        hereLogger.info(`${++this.numTriggers} triggers so far, state:`, this.statusDict());
         let feedAgeInMinutes = this.mostRecentHomeTootAgeInSeconds();
         if (feedAgeInMinutes) feedAgeInMinutes /= 60;
         const maxAgeMinutes = config.minTrendingMinutesUntilStale();
 
         if (isQuickMode && feedAgeInMinutes && feedAgeInMinutes < maxAgeMinutes && this.numTriggers <= 1) {
-            loggers[LogPrefix.TRIGGER_FEED_UPDATE].debug(`QUICK_MODE Feed's ${feedAgeInMinutes.toFixed(0)}s old, skipping`);
+            hereLogger.debug(`QUICK_MODE Feed's ${feedAgeInMinutes.toFixed(0)}s old, skipping`);
             // Needs to be called to update the feed in the app
             ScorerCache.prepareScorers().then((_t) => this.filterFeedAndSetInApp());
             return true;
@@ -631,8 +630,9 @@ class TheAlgorithm {
 
     // Do some final cleanup and scoring operations on the feed.
     private async finishFeedUpdate(): Promise<void> {
-        const hereLogger = loggers[LogPrefix.FINISH_FEED_UPDATE];
-        this.loadingStatus = FINALIZING_SCORES_MSG;
+        const action = LogAction.FINISH_FEED_UPDATE;
+        const hereLogger = loggers[action];
+        this.loadingStatus = config.locale.messages[action];
 
         // Now that all data has arrived go back over the feed and do the slow calculations of trendingLinks etc.
         hereLogger.debug(`${this.loadingStatus}...`);
@@ -714,45 +714,35 @@ class TheAlgorithm {
     };
 
     // Throws an error if the feed is loading, otherwise lock the mutex and set the loadStartedAt timestamp.
-    private async lockLoadingMutex(logPrefix: LogPrefix): Promise<void> {
-        loggers[logPrefix].log(`called, state:`, this.statusDict());
+    private async startAction(logPrefix: LoadAction): Promise<void> {
+        const hereLogger = loggers[logPrefix];
+        const status = config.locale.messages[logPrefix];
+        hereLogger.log(`called, state:`, this.statusDict());
 
         if (this.isLoading) {
-            loggers[logPrefix].warn(`Load in progress already!`, this.statusDict());
-            throw new Error(GET_FEED_BUSY_MSG);
+            hereLogger.warn(`Load in progress already!`, this.statusDict());
+            throw new Error(config.locale.messages.isBusy);
         }
 
         this.loadStartedAt = new Date();
         this._releaseLoadingMutex = await lockExecution(this.loadingMutex, logger);
-
-        if (logPrefix in LOADING_STATUS_MSGS) {
-            this.loadingStatus = LOADING_STATUS_MSGS[logPrefix as LogPrefix]!;
-        } else if (!this.feed.length) {
-            this.loadingStatus = INITIAL_LOAD_STATUS;
-        } else if (this.homeFeed.length > 0) {
-            const mostRecentAt = this.mostRecentHomeTootAt();
-            this.loadingStatus = `Loading new toots` + optionalSuffix(mostRecentAt, t => `since ${timeString(t)}`);
-        } else {
-            this.loadingStatus = `Loading more toots (retrieved ${this.feed.length.toLocaleString()} toots so far)`;
-        }
+        this.loadingStatus = (typeof status === 'string') ? status : status(this.feed, this.mostRecentHomeTootAt());
     }
 
     // Merge newToots into this.feed, score, and filter the feed.
     // NOTE: Don't call this directly! Use lockedMergeTootsToFeed() instead.
     private async mergeTootsToFeed(newToots: Toot[], inLogger: Logger): Promise<void> {
-        const startedAt = new Date();
+        const hereLogger = inLogger.tempLogger('mergeTootsToFeed');
         const numTootsBefore = this.feed.length;
-        this.feed = Toot.dedupeToots([...this.feed, ...newToots], inLogger.tempLogger('mergeTootsToFeed'));
+        const startedAt = new Date();
+
+        this.feed = Toot.dedupeToots([...this.feed, ...newToots], hereLogger);
         await updateBooleanFilterOptions(this.filters, this.feed);
         await this.scoreAndFilterFeed();
-        inLogger.logTelemetry(`merged ${newToots.length} new toots into ${numTootsBefore} timeline toots`, startedAt);
 
-        if (this.homeFeed.length > 0) {
-            const mostRecentAt = this.mostRecentHomeTootAt();
-            this.loadingStatus = `Loading new toots` + optionalSuffix(mostRecentAt, t => `since ${timeString(t)}`);
-        } else {
-            this.loadingStatus = `Loading more toots (retrieved ${this.feed.length.toLocaleString()} toots so far)`;
-        }
+        const statusMsgFxn = config.locale.messages[LoadAction.FEED_UPDATE];
+        this.loadingStatus = statusMsgFxn(this.feed, this.mostRecentHomeTootAt());
+        hereLogger.logTelemetry(`Merged ${newToots.length} new toots into ${numTootsBefore} timeline toots`, startedAt);
     }
 
     // Recompute the scorers' computations based on user history etc. and trigger a rescore of the feed
@@ -763,7 +753,7 @@ class TheAlgorithm {
     }
 
     // Release the loading mutex and reset the loading state variables.
-    private releaseLoadingMutex(logPrefix: LogPrefix): void {
+    private releaseLoadingMutex(logPrefix: LoadAction): void {
         this.loadingStatus = null;
 
         if (this._releaseLoadingMutex) {
@@ -835,7 +825,9 @@ class TheAlgorithm {
 };
 
 
-export default TheAlgorithm;
+// Some strings we want to export from the config
+const GET_FEED_BUSY_MSG = config.locale.messages[LoadAction.IS_BUSY];
+const READY_TO_LOAD_MSG = config.locale.messages[LogAction.INITIAL_LOADING_STATUS];
 
 // Export types and constants needed by apps using this package
 export {