+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # NEXT RELEASE
 
+### v0.74.3
+* Create config.locale.messages for potentially user facing string configuration
+* Reset trendingData when reset() is called
+* Refactor out FediverseCacheKey enum
+
 ### v0.74.2
-* Rename TagTootsType enum to TagTootsCategory
+* Rename `TagTootsType` enum to `TagTootsCategory`
 
 ### v0.74.1
 * Only do the followed hashtag additional scan in finishFeedUpdate(), not for each fetch because it's slow